+++ src/position.cpp
@@ -1,13 +1,15 @@
 #include "position.h"
 
-#include <cassert>
-
 #include "attacks.h"
+#include "zobrist.h"
+
+#include <cassert>
 
 Position Position::make_move(const Move& move) const
 {
     Position new_pos = *this;
     new_pos.Turn = !new_pos.Turn;
+    new_pos.Key ^= Zobrist.turn;
 
     if(move == no_move)
     {
@@ -21,6 +23,7 @@ Position Position::make_move(const Move& move) const
     new_pos.Squares[move.To] = move.Turn;
     new_pos.Bitboards[move.Turn] |= get_bitboard(move.To);
     new_pos.Bitboards[Pieces::Empty] &= ~get_bitboard(move.To);
+    new_pos.Key ^= Zobrist.squares[move.Turn][move.To];
 
     // FROM
     if(move.From != no_square)
@@ -30,6 +33,7 @@ Position Position::make_move(const Move& move) const
         new_pos.Squares[move.From] = Pieces::Empty;
         new_pos.Bitboards[move.Turn] &= ~get_bitboard(move.From);
         new_pos.Bitboards[Pieces::Empty] |= get_bitboard(move.From);
+        new_pos.Key ^= Zobrist.squares[move.Turn][move.From];
     }
 
     // CAPTURE
@@ -40,8 +44,10 @@ Position Position::make_move(const Move& move) const
     {
         const auto attacked_square = pop_lsb(attacked);
         new_pos.Squares[attacked_square] = move.Turn;
+        new_pos.Key ^= Zobrist.squares[move.Turn][attacked_square];
+        new_pos.Key ^= Zobrist.squares[!move.Turn][attacked_square];
     }
-    
+
     return new_pos;
 }